resumate/backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pdfplumber==0.10.3
pillow==10.2.0
pytesseract==0.3.10
python-dotenv==1.0.1
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
flask-jwt-extended==4.6.0
python-jose==3.3.0